[lit-element] Refactor LitElement (#1237)
Refactor LitElement and UpdatingElement

LitElement
* removes all polyfill code and places in `lit-element-polyfill` which is an in place patch and must be loaded separately (preserves support for styling shadowRoot without adoptedStylesheets).
* Includes caching of CSSResult (adopted from https://github.com/Polymer/lit-element/pull/952)
* `finalizeStyles` is now the main entry point for gathering element styles and can be overridden to customize. It is called once per class in `finalize`.
* `adoptStyles` is now always called, previously was only called if renderRoot was a shadowRoot.
* [lit-element] Refactor UpdatingElement (#1240)

UpdatingElement
* Streamlines some internal private API
* renamed `_getUpdateComplete` to `getUpdateComplete`
* when a reflecting property is undefined, its attribute is removed (previously it was unchanged)
* dirty check in `attributeChangedCallback` removed (requested by user)
* Fixes https://github.com/Polymer/lit-element/issues/965
* Additional minor refactoring
* Also fixes an issue where the default fromAttribute function was not run if only a toAttribute function was set.
* Reduces code size for instanceProperties
* Separates decorators into individual modules
* Also separates decorators tests.
* Moved tests around so decorators are only used in tests for those decorators.
