+++ packages/lit-element/package.json
@@ -27,8 +27,8 @@
     "build": "tsc",
     "build:watch": "tsc --watch",
     "build:babel-test": "babel src/test/lib/decorators_test.ts --out-file test/lib/decorators-babel_test.js",
-    "test": "wtr test/*test.js --node-resolve --root-dir ../../",
-    "test:watch": "wtr test/*test.js --node-resolve --root-dir ../../ --watch",
+    "test": "wtr test/*_test.js test/**/*_test.js --node-resolve --root-dir ../../",
+    "test:watch": "wtr test/*_test.js test/**/*_test.js --node-resolve --root-dir ../../ --watch",
     "checksize": "rollup -c ; cat lit-element.bundled.js | wc -c ; rm lit-element.bundled.js",
     "format": "prettier src/* --write",
     "lint": "tslint --project ./",
