+++ packages/lit-element/src/lib/css-tag.ts
@@ -21,9 +21,8 @@ export const supportsAdoptingStyleSheets =
 const constructionToken = Symbol();
 
 export class CSSResult {
-  _styleSheet?: CSSStyleSheet | null;
-
   readonly cssText: string;
+  private _styleSheet?: CSSStyleSheet;
 
   constructor(cssText: string, safeToken: symbol) {
     if (safeToken !== constructionToken) {
@@ -31,22 +30,17 @@ export class CSSResult {
         'CSSResult is not constructable. Use `unsafeCSS` or `css` instead.'
       );
     }
-
     this.cssText = cssText;
   }
 
   // Note, this is a getter so that it's lazy. In practice, this means
   // stylesheets are not created until the first element instance is made.
-  get styleSheet(): CSSStyleSheet | null {
-    if (this._styleSheet === undefined) {
-      // Note, if `supportsAdoptingStyleSheets` is true then we assume
-      // CSSStyleSheet is constructable.
-      if (supportsAdoptingStyleSheets) {
-        this._styleSheet = new CSSStyleSheet();
-        this._styleSheet.replaceSync(this.cssText);
-      } else {
-        this._styleSheet = null;
-      }
+  get styleSheet(): CSSStyleSheet | undefined {
+    // Note, if `supportsAdoptingStyleSheets` is true then we assume
+    // CSSStyleSheet is constructable.
+    if (supportsAdoptingStyleSheets && this._styleSheet === undefined) {
+      this._styleSheet = new CSSStyleSheet();
+      this._styleSheet.replaceSync(this.cssText);
     }
     return this._styleSheet;
   }
@@ -80,6 +74,8 @@ const textFromCSSResult = (value: CSSResult | number) => {
   }
 };
 
+const cssResultCache = new Map<string, CSSResult>();
+
 /**
  * Template tag which which can be used with LitElement's [[LitElement.styles |
  * `styles`]] property to set element styles. For security reasons, only literal
@@ -94,5 +90,12 @@ export const css = (
     (acc, v, idx) => acc + textFromCSSResult(v) + strings[idx + 1],
     strings[0]
   );
-  return new CSSResult(cssText, constructionToken);
+  let result = cssResultCache.get(cssText);
+  if (result === undefined) {
+    cssResultCache.set(
+      cssText,
+      (result = new CSSResult(cssText, constructionToken))
+    );
+  }
+  return result;
 };
