+++ packages/lit-element/src/test/lit-element_test.ts
@@ -13,7 +13,6 @@
  */
 
 import {html, LitElement} from '../lib/lit-element.js';
-import {property} from '../lib/decorators.js';
 import {stripExpressionComments} from 'lit-html/development/test/test-utils/strip-markers.js';
 import {generateElementName} from './test-helpers.js';
 import {assert} from '@esm-bundle/chai';
@@ -96,8 +95,7 @@ suite('LitElement', () => {
     assert.equal(stripExpressionComments(el.shadowRoot!.innerHTML), rendered);
   });
 
-  // TODO(sorvell): Renable when lit-next supports this.
-  test.skip('updates/renders attributes, properties, and event listeners via `lit-html`', async () => {
+  test('updates/renders attributes, properties, and event listeners via `lit-html`', async () => {
     class E extends LitElement {
       _event?: Event;
 
@@ -122,8 +120,7 @@ suite('LitElement', () => {
     assert.equal(el._event, e);
   });
 
-  // TODO(sorvell): Renable when lit-next supports this.
-  test.skip('event listeners are invoked with the right `this` value', async () => {
+  test('event listeners are invoked with the right `this` value', async () => {
     class E extends LitElement {
       event?: Event;
 
@@ -145,8 +142,7 @@ suite('LitElement', () => {
     assert.equal(el.event, event);
   });
 
-  // TODO(sorvell): Renable when lit-next supports this.
-  test.skip('can set properties and attributes on sub-element', async () => {
+  test('can set properties and attributes on sub-element', async () => {
     class E extends LitElement {
       static get properties() {
         return {foo: {}, attr: {}, bool: {type: Boolean}};
@@ -206,8 +202,7 @@ suite('LitElement', () => {
     assert.equal(window['litElementVersions'].length, 1);
   });
 
-  // TODO(sorvell): Renable when lit-next supports this.
-  test.skip('event fired during rendering element can trigger an update', async () => {
+  test('event fired during rendering element can trigger an update', async () => {
     class E extends LitElement {
       connectedCallback() {
         super.connectedCallback();
@@ -245,7 +240,8 @@ suite('LitElement', () => {
   test('exceptions in `render` throw but do not prevent further updates', async () => {
     let shouldThrow = false;
     class A extends LitElement {
-      @property() foo = 5;
+      static properties = {foo: {}};
+      foo = 5;
       updatedFoo = 0;
 
       render() {
